Seperate distribute and the normaldistribution funciton
