+++ production.js
@@ -1,0 +1,15 @@
+const distribution = require("./distribute");
+
+class Production{
+
+    constructor(){
+        this.dist = new distribution;
+        this.
+
+
+    }
+
+    prodAvg(){
+    }
+
+}
